kill utils.js, reduce (set|get)(Local|Session) to just lbry.js
