+++ ui/js/reducers/user.js
@@ -1,5 +1,4 @@
 import * as types from "constants/action_types";
-import { getLocal } from "utils";
 
 const reducers = {};
